+++ tests/phpunit/Integration/TestCase.php
@@ -1,7 +1,7 @@
 <?php
 declare ( strict_types = 1 );
 
-namespace PixelgradeLT\PartTemplate\Tests\Integration;
+namespace Pressody\PartTemplate\Tests\Integration;
 
 use WP_UnitTestCase;