+++ tests/phpunit/Unit/TestCase.php
@@ -1,7 +1,7 @@
 <?php
 declare ( strict_types = 1 );
 
-namespace PixelgradeLT\PartTemplate\Tests\Unit;
+namespace Pressody\PartTemplate\Tests\Unit;
 
 use Brain\Monkey;
 use PHPUnit\Framework\TestCase as PHPUnitTestCase;